modified the consultant assignment to have dynamic state and district , and level as well

+++ src/app/api/consultants/assign/route.js
@@ -3,7 +3,14 @@ import { NextResponse } from "next/server";
 
 export async function POST(request) {
   try {
-    const { consultant, finalSectors, finalDomains } = await request.json();
+    const {
+      consultant,
+      finalSectors,
+      finalDomains,
+      finalStates,
+      finalDistricts,
+      assignedLevel,
+    } = await request.json();
     // console.log("backendddddd", consultant, finalSectors, finalDomains);
 
     // Check if the consultant exists
@@ -28,6 +35,16 @@ export async function POST(request) {
         { status: 400 }
       );
     }
+    // Check if both finalStates and finalDistricts have length greater than 0
+    if (finalStates.length === 0 || finalDistricts.length === 0) {
+      return NextResponse.json(
+        {
+          message:
+            "Both states and districts should have at least one assignment.",
+        },
+        { status: 400 }
+      );
+    }
 
     /*** 1️⃣ Delete Existing Assignments ***/
     await db.consultantAssignedSectors.deleteMany({
@@ -35,6 +52,12 @@ export async function POST(request) {
     });
 
     await db.consultantAssignedDomains.deleteMany({
+      where: { consultantId: consultant },
+    });
+    await db.ConsultantAssignedStates.deleteMany({
+      where: { consultantId: consultant },
+    });
+    await db.ConsultantAssignedDistricts.deleteMany({
       where: { consultantId: consultant },
     });
 
@@ -58,8 +81,33 @@ export async function POST(request) {
       });
     }
 
+    /*** 4 Insert New State Assignments ***/
+    if (finalStates.length > 0) {
+      await db.ConsultantAssignedStates.createMany({
+        data: finalStates.map((stateId) => ({
+          consultantId: consultant,
+          stateId,
+        })),
+      });
+    }
+    /*** 3️⃣ Insert New District Assignments ***/
+    if (finalDistricts.length > 0) {
+      await db.ConsultantAssignedDistricts.createMany({
+        data: finalDistricts.map((districtId) => ({
+          consultantId: consultant,
+          districtId,
+        })),
+      });
+    }
+
+    /*** 6️⃣ Update Consultant Profile's assignedLevel ***/
+    await db.consultantProfile.update({
+      where: { id: consultant },
+      data: { assignedLevel }, // Update only assignedLevel
+    });
+
     return NextResponse.json(
-      { message: "Sectors and Domains Assigned Successfully" },
+      { message: "Sectors, Domains, Location Assigned Successfully" },
       { status: 201 }
     );
   } catch (error) {
